requirements.txt
beautifulsoup4==4.12.2
newspaper3k==0.2.8
nltk==3.7
pandas==1.5.2
Requests==2.31.0
scikit_learn==1.3.1
tqdm==4.64.1
